Cleanup e2e tests
